% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{as_bb}
\alias{as_bb}
\title{Extract a bb vector of [xmin, ymin, xmax, ymax] from numeric vector or sf object}
\usage{
as_bb(x)
}
\arguments{
\item{x}{object accepted by st_bbox()}
}
\value{
bb vector of [xmin, ymin, xmax, ymax]
}
\description{
Extract a bb vector of [xmin, ymin, xmax, ymax] from numeric vector or sf object
}
