% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{as_POINT}
\alias{as_POINT}
\title{Convert a data frame (or tibble) to an sf POINT object}
\usage{
as_POINT(x = read_gom(), crs = 4326, dims = "xy", ...)
}
\arguments{
\item{x}{data frame or tibble}

\item{crs}{the CRS of the location coordinates}

\item{dims}{character, one of 'xy', 'xyz' 'xyt' or 'xyzt'}

\item{...}{other arguments passed to \code{\link[sf]{st_as_sf}}}
}
\value{
sf POINT object
}
\description{
Convert a data frame (or tibble) to an sf POINT object
}
