% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{as_st_bbox}
\alias{as_st_bbox}
\title{Construct a bbox from a 'bb' object or [xmin, ymin, xmax, ymax] numeric vector}
\usage{
as_st_bbox(x, crs = 4326)
}
\arguments{
\item{x}{object accepted by \code{\link{as_bb}}}

\item{crs}{coordinate reference system, such as '4326'}
}
\value{
\code{\link[sf]{st_bbox}}
}
\description{
Construct a bbox from a 'bb' object or [xmin, ymin, xmax, ymax] numeric vector
}
