% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{assert_time}
\alias{assert_time}
\title{Assert the the input has a valid time (\code{t}) variable}
\usage{
assert_time(x)
}
\arguments{
\item{x}{xyzt tibble}
}
\value{
the input tibble
}
\description{
Assert the the input has a valid time (\code{t}) variable
}
