% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{assert_x}
\alias{assert_x}
\title{Assert the input has a valid \code{x} (lon) variable}
\usage{
assert_x(x)
}
\arguments{
\item{x}{xyzt tibble}
}
\value{
the input tibble
}
\description{
Assert the input has a valid \code{x} (lon) variable
}
