% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{assert_y}
\alias{assert_y}
\title{Assert the input has a valid \code{y} (lat) variable}
\usage{
assert_y(x)
}
\arguments{
\item{x}{xyzt tibble}
}
\value{
the input tibble
}
\description{
Assert the input has a valid \code{y} (lat) variable
}
