% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{assert_z}
\alias{assert_z}
\title{Assert the input has a valid \code{z} (depth/altitude) variable}
\usage{
assert_z(x)
}
\arguments{
\item{x}{xyzt tibble}
}
\value{
the input tibble
}
\description{
Assert the input has a valid \code{z} (depth/altitude) variable
}
