% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_straddles}
\alias{bb_straddles}
\title{Test if a bounding box straddles a longitude}
\usage{
bb_straddles(bb = c(-170, -60, 50, 60), at = 0)
}
\arguments{
\item{bb}{numeric, 4 element bounding box of [xmin, ymin, xmax, ymax] coordinates
or an object accepted by \code{\link{as_bb}}}

\item{at}{numeric, longitude to test straddle-ness}
}
\value{
logical
}
\description{
Test if a bounding box straddles a longitude
}
