% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_to_180}
\alias{bb_to_180}
\title{Convert bounding box [0,360] longitudes to [-180, 180]}
\usage{
bb_to_180(x)
}
\arguments{
\item{x}{numeric bounding box vector, no check is done for being withing 0,360 range}
}
\value{
numeric bounding box vector
}
\description{
Bounding boxes are 4 element vectors of [xmin, ymin, xmax, ymax]
}
