% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_to_360}
\alias{bb_to_360}
\title{Convert [-180,180] bounding box longitudes to [0,360]}
\usage{
bb_to_360(x)
}
\arguments{
\item{x}{numeric bounding box vector, no check is done for being withing 0,360 range}
}
\value{
numeric bounding box vector
}
\description{
Bounding boxes are 4 element vectors of [xmin, ymin, xmax, ymax]
}
