% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract.default}
\alias{extract.stars}
\alias{extract.ncdf4}
\title{extract generic}
\usage{
extract(x, ...)

\method{extract}{default}(x, y = NULL, ...)

\method{extract}{stars}(x, y = NULL, verbose = FALSE, ...)

\method{extract}{ncdf4}(x, y = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{ncdf4} object}

\item{...}{arguments to be passed to source data \code{extract} method}

\item{y}{\code{sf} object}

\item{verbose}{logical, output helpful messages?}
}
\value{
data frame of covariates (point, or raster)
}
\description{
extract generic
}
\section{Methods (by class)}{
\itemize{
\item \code{extract(default)}: Extract data from a NCDF4 object

\item \code{extract(stars)}: Extract from a stars object using any sf object

\item \code{extract(ncdf4)}: Extract from a NCDF4 object using any sf object

}}
