% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{get_geometry_dimension}
\alias{get_geometry_dimension}
\title{Get geometry dimension code}
\usage{
get_geometry_dimension(x)
}
\arguments{
\item{x}{sf or sfc object}
}
\value{
character vector such as "XY" or "XYZ"
}
\description{
Get geometry dimension code
}
