% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{get_geometry_type}
\alias{get_geometry_type}
\title{Get geometry type code}
\usage{
get_geometry_type(x)
}
\arguments{
\item{x}{sf or sfc object}
}
\value{
character vector such as "POINT" or "POLYGON"
}
\description{
Get geometry type code
}
