% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{is_source}
\alias{is_source}
\title{Given a URL or ncdf object test if the data matches the 
specified data source pattern}
\usage{
is_source(x, pattern = "hycom", fixed = TRUE)
}
\arguments{
\item{x}{character or NCDF4 object}

\item{pattern}{charcater, fixed or regex pattern to match}

\item{fixed}{logical, is the \code{pattern} argument
a fixed pattern or regular expression}
}
\value{
logical where TRUE finds a match
}
\description{
Given a URL or ncdf object test if the data matches the 
specified data source pattern
}
