% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{POSIX_epoch}
\alias{POSIX_epoch}
\title{Retrieve the POSIX epoch}
\usage{
POSIX_epoch(x = "1970-01-01 00:00:00", tz = "UTC", ...)
}
\arguments{
\item{x}{character date and time of epoch}

\item{tz}{character, time zone}

\item{...}{other arguments for \code{\link[base]{as.POSIXct}}}
}
\value{
POSIXct time
}
\description{
Retrieve the POSIX epoch
}
