% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{read_calcofi}
\alias{read_calcofi}
\title{Read an example point dataset for the CalCOFI Sampling Stations (since 1984)}
\usage{
read_calcofi(
  url = "https://calcofi.org/downloads/maps/CalCOFIStationOrder.csv"
)
}
\arguments{
\item{url}{character string url to data}
}
\value{
a tibble of CalCOFI points
}
\description{
Read an example point dataset for the CalCOFI Sampling Stations (since 1984)
}
