% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{read_gom}
\alias{read_gom}
\title{Read an example point dataset for Gulf of Maine}
\usage{
read_gom(
  filename = system.file(file.path("extdata", "gom.csv"), package = "xyzt")
)
}
\arguments{
\item{filename}{character, the file to read}
}
\value{
tibble
}
\description{
Read an example point dataset for Gulf of Maine
}
