% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{read_sab}
\alias{read_sab}
\title{Read an example point dataset for South Atlantic Bight}
\usage{
read_sab(
  filename = system.file(file.path("extdata", "sab.csv"), package = "xyzt")
)
}
\arguments{
\item{filename}{character, the file to read}
}
\value{
tibble
}
\description{
Read an example point dataset for South Atlantic Bight
}
