% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{var_lut}
\alias{var_lut}
\title{Retrieve a series of allowed variable name associations}
\usage{
var_lut()
}
\value{
list of variable name associations
}
\description{
Retrieve a series of allowed variable name associations
}
