% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzt.R
\name{as_BBOX}
\alias{as_BBOX}
\title{Convert any object to a bounding box POLYGON.  
Note that attributes (variables) are dropped.}
\usage{
as_BBOX(x)
}
\arguments{
\item{x}{sf object}
}
\value{
sf object with one POLYGON record
}
\description{
Convert any object to a bounding box POLYGON.  
Note that attributes (variables) are dropped.
}
